% Auto-generated: do not edit by hand
\name{''CustomTransformAccordion}

\alias{''CustomTransformAccordion}

\title{CustomTransformAccordion component}

\description{

}

\usage{
''CustomTransformAccordion(children=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

